fix loader
