+++ webpack/loaders/manifest-loader.ts
@@ -1,17 +1,23 @@
+import fs from 'fs';
+
 import webpack from 'webpack';
 
-const commentRegex = /(?<=^(?:[^"/]*(?:|"[^"]*"))[^"/]*(?:|sync\s*\/[^/]*\/[^"/]*))#.*$/i;
-
-export default function(this: webpack.LoaderContext<never>, content: string): string {
+export default function(this: webpack.LoaderContext<never>, _content: string): string {
   this.cacheable(true);
-  let ret = '';
+  const contents = fs.readFileSync(this.resourcePath).toString();
+  const lines = contents.trim().split(/\s*[\r\n]+\s*/);
 
-  content.split(/\r?\n/).forEach((_line) => {
-    const line = _line.replace(commentRegex, '').trim();
-    if (!line)
-      return;
-    ret += line + '\r\n';
+  let importStr = '';
+  let outputStr = 'export default {';
+  lines.forEach((rawName, fileIdx) => {
+    // normalize filepaths between windows / unix
+    const name = rawName.replace(/\\/g, '/').replace(/^\//, '');
+    // Use static imports instead of dynamic ones to put files in the bundle.
+    const fileVar = `file${fileIdx}`;
+    importStr += `import ${fileVar} from './${name}';\n`;
+    outputStr += `'${name}': ${fileVar},`;
   });
+  outputStr += '};';
 
-  return ret;
+  return `${importStr}\n${outputStr}`;
 }
